+++ LessonCh006/Models/ChiefAccountant.cs
@@ -27,7 +27,7 @@
         /// <param name="age"> Возраст </param>
         /// <param name="salary"> Зарплата </param>
         /// <param name="jobTitle"> Название должности </param>
-        public ChiefAccountant(int id, string name, string surname, long age, double salary, string jobTitle) : base(id, name, surname, age, salary, jobTitle)
+        public ChiefAccountant(ulong id, string name, string surname, long age, double salary, string jobTitle) : base(id, name, surname, age, salary, jobTitle)
         {
             EmployeePosition = EmployeePosition.ChiefAccountant;
         }
